pyproject.toml
[project]
name = "pr-livro-loja"
version = "0.1.0"
description = ""
authors = [
    {name = "redssy",email = "redlemagic@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "faker (>=37.0.2,<38.0.0)",
    "pytest-django (>=4.10.0,<5.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"